feat: init git + classutils gen + API-only mode

+++ templates/tilt.go
@@ -8,11 +8,14 @@ analytics_settings(enable=False)
 
 # custom config
 config.define_bool('use-local', usage='use this to include our own libraries in dev mode (like goaldr)')
+config.define_bool('api-only', usage='use this to only work with the API part of this project')
 cfg = config.parse()
 useLocalDeps = cfg.get('use-local')
+apiOnly = cfg.get('api-only')
 
 # describing the deployment of all the backend services, and their configuration
-if useLocalDeps:
+if useLocalDeps or apiOnly:
+  # when working with local dependencies, or in API-only mode, we do not dockerise the web app
   k8s_yaml(kustomize('{{.Deploying.Dir}}/overlays/local'))
 else:
   # when working with no use of local deps, the base config is enough
@@ -27,7 +30,7 @@ local_resource(
     # taking into account the dependencies
     deps  =['{{.API.SrcDir}}', '../goald'],
     # the API config is also ignored here, because Aldev is already watching it
-    ignore=['{{.API.SrcDir}}/go.sum', '{{.API.SrcDir}}/_generated', '{{.API.SrcDir}}/**/*--utils.go', '{{.API.Config}}'],
+    ignore=['{{.API.SrcDir}}/go.sum', '{{.API.SrcDir}}/_generated', '{{.API.SrcDir}}/**/*/classutils', '{{.API.Config}}'],
     )
 
 # describing the containers for the backend - cf https://docs.tilt.dev/extensions.html
@@ -51,8 +54,6 @@ k8s_resource('{{.AppName}}-api-depl', resource_deps=['{{.AppName}}-api-compile']
 apiHost = str(local(echo_off=True, command="kubectl get services --namespace kube-system "+
   "-o jsonpath='{.items[?(@.spec.type==\"LoadBalancer\")].status.loadBalancer.ingress[0].ip}'"))
 
-# --- WEB part ----------------------------------------------------------------
-
 # since the load balancer is on host 'apiHost' and not 'localhost', we can reach it by sending calls
 # to localhost:apiPort - from Windows for instance - and proxying them inside Linux to apiHost:apiPort
 local_resource(
@@ -62,42 +63,47 @@ local_resource(
     resource_deps=['{{.AppName}}-api-depl'],
     )
 
-# if we're developing libraries along with this project, then:
-if useLocalDeps:
-  webAppEnvVars = "WEB_API_URL=http://localhost:{{.API.Port}}"
-  {{range .Web.EnvVars}}webAppEnvVars += " {{.Name}}={{.Value}}"
-  {{end}}
-  # locally running Vite's dev server - no need to containerize this for now
-  local_resource(
-      name         ='{{.AppName}}-vite-serve',
-      dir          ='{{.Web.SrcDir}}',
-      cmd          ='rm -fr node_modules/.vite && npm i --force',
-      deps         =['{{.Web.SrcDir}}/package.json'],
-      serve_cmd    =webAppEnvVars + ' npm run dev',
-      serve_dir    ='{{.Web.SrcDir}}',
-      resource_deps=['{{.AppName}}-lb-proxy'],
-      )
-else:
-  docker_build(
-    '{{.AppName}}-web-image',
-    context='.',
-    dockerfile='./{{.Deploying.Dir}}/docker/{{.AppName}}-local-web-docker',
-    only=['{{.Web.SrcDir}}/'],
-    ignore=['{{.Web.SrcDir}}/dist/'],
-    live_update=[
-        fall_back_on('{{.Web.SrcDir}}/vite.config.js'),
-        sync('{{.Web.SrcDir}}/', '/web/'),
-        run(
-            'npm install --force',
-            trigger=['{{.Web.SrcDir}}/package.json', '{{.Web.SrcDir}}/package-json.lock']
-        )
-    ]
-  )
+# --- WEB part ----------------------------------------------------------------
 
-  k8s_resource(
-      '{{.AppName}}-web-depl',
-      port_forwards='{{.Web.Port}}:5173', # 5173 is the port Vite listens on in the container
-      resource_deps=['{{.AppName}}-api-depl'],
-      # labels=['frontend']
-  )
+if not apiOnly:
+  # if we're developing libraries along with this project, then:
+  if useLocalDeps:
+    webAppEnvVars = "WEB_API_URL=http://localhost:{{.API.Port}}"
+    {{range .Web.EnvVars}}webAppEnvVars += " {{.Name}}={{.Value}}"
+    {{end}}
+    # locally running Vite's dev server - no need to containerize this for now
+    local_resource(
+        name         ='{{.AppName}}-vite-serve',
+        dir          ='{{.Web.SrcDir}}',
+        cmd          ='rm -fr node_modules/.vite && npm i --force',
+        deps         =['{{.Web.SrcDir}}/package.json'],
+        serve_cmd    =webAppEnvVars + ' npm run dev',
+        serve_dir    ='{{.Web.SrcDir}}',
+        resource_deps=['{{.AppName}}-lb-proxy'],
+        )
+  else:
+    docker_build(
+      '{{.AppName}}-web-image',
+      context='.',
+      dockerfile='./{{.Deploying.Dir}}/docker/{{.AppName}}-local-web-docker',
+      only=['{{.Web.SrcDir}}/'],
+      ignore=['{{.Web.SrcDir}}/dist/'],
+      live_update=[
+          fall_back_on('{{.Web.SrcDir}}/vite.config.js'),
+          sync('{{.Web.SrcDir}}/', '/web/'),
+          run(
+              'npm install --force',
+              trigger=['{{.Web.SrcDir}}/package.json', '{{.Web.SrcDir}}/package-json.lock']
+          )
+      ]
+    )
+
+    k8s_resource(
+        '{{.AppName}}-web-depl',
+        port_forwards='{{.Web.Port}}:5173', # 5173 is the port Vite listens on in the container
+        resource_deps=['{{.AppName}}-api-depl'],
+        # labels=['frontend']
+    )
+else:
+  print("\nAPI only mode")
 ` // end of template
